Fix Printing Structure After It Has Run (#1437)
