+++ tests/unit/structures/test_agent.py
@@ -255,50 +255,6 @@ class TestAgent:
         assert len(agent.task_outputs) == 1
         assert agent.task_outputs[task.id] == task.output
 
-    def test_to_dict(self):
-        task = PromptTask("test prompt")
-        agent = Agent(prompt_driver=MockPromptDriver())
-        agent.add_task(task)
-        expected_agent_dict = {
-            "type": "Agent",
-            "id": agent.id,
-            "tasks": [
-                {
-                    "type": agent.tasks[0].type,
-                    "id": agent.tasks[0].id,
-                    "state": str(agent.tasks[0].state),
-                    "parent_ids": agent.tasks[0].parent_ids,
-                    "child_ids": agent.tasks[0].child_ids,
-                    "max_meta_memory_entries": agent.tasks[0].max_meta_memory_entries,
-                    "context": agent.tasks[0].context,
-                }
-            ],
-            "conversation_memory": {
-                "type": agent.conversation_memory.type,
-                "runs": agent.conversation_memory.runs,
-                "meta": agent.conversation_memory.meta,
-                "max_runs": agent.conversation_memory.max_runs,
-            },
-        }
-        assert agent.to_dict() == expected_agent_dict
-
-    def test_from_dict(self):
-        task = PromptTask("test prompt")
-        agent = Agent(prompt_driver=MockPromptDriver())
-        agent.add_task(task)
-
-        serialized_agent = agent.to_dict()
-        assert isinstance(serialized_agent, dict)
-
-        deserialized_agent = Agent.from_dict(serialized_agent)
-        assert isinstance(deserialized_agent, Agent)
-
-        assert deserialized_agent.task_outputs[task.id] is None
-        deserialized_agent.run()
-
-        assert len(deserialized_agent.task_outputs) == 1
-        assert deserialized_agent.task_outputs[task.id].value == "mock output"
-
     def test_runnable_mixin(self):
         mock_on_before_run = Mock()
         mock_after_run = Mock()
